+++ app/join/components/JoinForm.tsx
@@ -1,16 +1,11 @@
 "use client";
 import { useState, useRef, ChangeEventHandler } from "react";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { zodJoinSchema } from "../zodJoinSchema";
+import { FieldErrors, UseFormRegister } from "react-hook-form";
 import Image from "next/image";
 import visible from "../../../public/join/visible.png";
-import Button from "@/components/Button";
 import plus from "../../../public/join/plus.png";
-import Link from "next/link";
-import { locationStore } from "@/store/locationStore";
 import JoinInputField from "./JoinInputFiled";
-import { useSignup } from "../lib/useSignup";
+
 export interface JoinFormData {
   email: string;
   name: string;
@@ -19,33 +14,27 @@ export interface JoinFormData {
   address: string;
   sortRole: boolean;
 }
-
-export default function JoinForm() {
-  const mutation = useSignup();
-  const { location, setLocation } = locationStore();
+interface JoinFormProps {
+  register: UseFormRegister<JoinFormData>;
+  errors: FieldErrors<JoinFormData>;
+  clickSubmit: boolean;
+  location: string;
+}
+export default function JoinForm({
+  register,
+  errors,
+  location,
+  clickSubmit,
+}: JoinFormProps) {
   const [showPassword, setShowPassword] = useState<boolean>(false);
   const imageRef = useRef<HTMLInputElement>(null);
   const [selectedImage, setSelectedImage] = useState<string>();
-  const [clickSubmit, setClickSubmit] = useState<boolean>(false);
+
   const onClickPictureButton = () => {
     imageRef.current?.click();
   };
   const togglePasswordVisibility = () => {
     setShowPassword(!showPassword);
-  };
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<JoinFormData>({
-    resolver: zodResolver(zodJoinSchema),
-  });
-  const onSubmit = (data: JoinFormData) => {
-    setClickSubmit(true);
-    if (location) {
-      const { email, name, password, sortRole } = data;
-      mutation.mutate({ address: location, email, name, password, sortRole });
-    }
   };
 
   const handleImageChange: ChangeEventHandler<HTMLInputElement> = (e) => {
@@ -57,6 +46,7 @@ export default function JoinForm() {
       };
     }
   };
+
   return (
     <div className="w-full">
       <div className="flex flex-row justify-center">
@@ -89,100 +79,90 @@ export default function JoinForm() {
         </div>
       </div>
 
-      <form
-        onSubmit={handleSubmit(onSubmit)}
-        className="flex flex-col mx-4 flex-1  my-5"
-      >
+      <JoinInputField
+        title="이메일"
+        placeholder="이메일을 입력하세요"
+        register={{ ...register("email") }}
+      />
+
+      {errors.email && (
+        <p className="text-red-600 font-bold">{errors.email.message}</p>
+      )}
+      <JoinInputField
+        title="이름"
+        placeholder="이름을 입력하세요"
+        register={{ ...register("name") }}
+      />
+
+      {errors.name && (
+        <p className="text-red-600 font-bold">{errors.name.message}</p>
+      )}
+
+      <div className="relative my-3">
         <JoinInputField
-          title="이메일"
-          placeholder="이메일을 입력하세요"
-          register={{ ...register("email") }}
+          title="비밀번호"
+          placeholder="비밀번호를 입력하세요"
+          register={{ ...register("password") }}
+          type={showPassword ? "text" : "password"}
+        />
+        <button
+          type="button"
+          onClick={togglePasswordVisibility}
+          className="absolute right-2 top-1/2"
+        >
+          <Image src={visible} alt="visible" width={20} height={20} />
+        </button>
+        {errors.password && (
+          <p className="text-red-600 font-bold">{errors.password.message}</p>
+        )}
+      </div>
+
+      <div className="relative my-3">
+        <JoinInputField
+          title="비밀번호 확인"
+          placeholder="비밀번호를 한번더 입력하세요."
+          register={{ ...register("passwordConfirmation") }}
+          type={showPassword ? "text" : "password"}
         />
 
-        {errors.email && (
-          <p className="text-red-600 font-bold">{errors.email.message}</p>
+        <button
+          type="button"
+          onClick={togglePasswordVisibility}
+          className="absolute right-2 top-1/2"
+        >
+          <Image src={visible} alt="visible" width={20} height={20} />
+        </button>
+        {errors.passwordConfirmation && (
+          <p className="text-red-600 font-bold">
+            {errors.passwordConfirmation.message}
+          </p>
         )}
-        <JoinInputField
-          title="이름"
-          placeholder="이름을 입력하세요"
-          register={{ ...register("name") }}
+      </div>
+
+      <label className="text-[#323232] flex flex-col font-bold my-1">
+        주소
+        <input
+          type="text"
+          placeholder="주소를 입력하세요"
+          className="bg-[#e4e6e7] rounded-[6px] py-2 font-normal"
+          value={location}
+          disabled
         />
-
-        {errors.name && (
-          <p className="text-red-600 font-bold">{errors.name.message}</p>
-        )}
-
-        <div className="relative my-3">
-          <JoinInputField
-            title="비밀번호"
-            placeholder="비밀번호를 입력하세요"
-            register={{ ...register("password") }}
-            type={showPassword ? "text" : "password"}
-          />
-          <button
-            type="button"
-            onClick={togglePasswordVisibility}
-            className="absolute right-2 top-1/2"
-          >
-            <Image src={visible} alt="visible" width={20} height={20} />
-          </button>
-          {errors.password && (
-            <p className="text-red-600 font-bold">{errors.password.message}</p>
-          )}
-        </div>
-
-        <div className="relative my-3">
-          <JoinInputField
-            title="비밀번호 확인"
-            placeholder="비밀번호를 한번더 입력하세요."
-            register={{ ...register("passwordConfirmation") }}
-            type={showPassword ? "text" : "password"}
-          />
-
-          <button
-            type="button"
-            onClick={togglePasswordVisibility}
-            className="absolute right-2 top-1/2"
-          >
-            <Image src={visible} alt="visible" width={20} height={20} />
-          </button>
-          {errors.passwordConfirmation && (
-            <p className="text-red-600 font-bold">
-              {errors.passwordConfirmation.message}
-            </p>
-          )}
-        </div>
-
-        <label className="text-[#323232] flex flex-col font-bold my-1">
-          주소
-          <input
-            type="text"
-            placeholder="주소를 입력하세요"
-            className="bg-[#e4e6e7] rounded-[6px] py-2 font-normal"
-            value={location}
-            disabled
-          />
+      </label>
+      {clickSubmit && location.length === 0 && (
+        <p className="text-red-600 font-bold">주소를 입력하세요</p>
+      )}
+      <div>
+        <input type="checkbox" id="isTrainer" {...register("sortRole")} />
+        <label htmlFor="isTrainer" className="text-[#323232]">
+          트레이너 입니다.
         </label>
-        {clickSubmit && location.length === 0 && (
-          <p className="text-red-600 font-bold">주소를 입력하세요</p>
-        )}
-        <div>
-          <input type="checkbox" id="isTrainer" {...register("sortRole")} />
-          <label htmlFor="isTrainer" className="text-[#323232]">
-            트레이너 입니다.
-          </label>
-        </div>
-        {errors.sortRole && (
-          <p className="text-red-600 font-bold">{errors.sortRole.message}</p>
-        )}
-        <Link href="../../join/locationModal">
-          <Button>주소찾기</Button>
-        </Link>
-        <Button type="submit" backgroundColor="bg-[#1C743C]" className="mt-2.5">
-          회원가입
-        </Button>
-        <div className=" pb-10" />
-      </form>
+      </div>
+      {errors.sortRole && (
+        <p className="text-red-600 font-bold">{errors.sortRole.message}</p>
+      )}
+
+      <div className=" pb-10" />
     </div>
   );
 }
